Implement checkout block support for Mollie blocks in the admin.
